[build] Fix libnl-genl link flags and size_t conversion warning
- Add -lnl-genl-3 -lnl-3 to all SAILIB definitions and test LDADD
  that link against libsaivs (syncd, saidiscovery, saisdkdump, tests,
  syncd/tests, vslib/tests) since libsaivs now depends on libnl-genl
  for the STEL genetlink sender.

- Fix -Werror=conversion in SwitchStateBaseStel.cpp: cast data_len
  from size_t to int for nla_put() third parameter.

Build verified: dpkg-buildpackage -Psyncd,vs,nopython2 passes in
sonic-slave-bookworm CI container with CI artifacts.

+++ vslib/SwitchStateBaseStel.cpp
@@ -111,7 +111,7 @@ struct GenlConnection
             return -1;
         }
 
-        int ret = nla_put(msg, SONIC_STEL_ATTR_IPFIX_DATA, data_len, data);
+        int ret = nla_put(msg, SONIC_STEL_ATTR_IPFIX_DATA, static_cast<int>(data_len), data);
         if (ret < 0)
         {
             SWSS_LOG_ERROR("nla_put() failed: %s", nl_geterror(ret));
